+++ eiam-20211201/src/Models/SetApplicationSsoConfigRequest/samlSsoConfig.php
@@ -5,6 +5,7 @@
 namespace AlibabaCloud\SDK\Eiam\V20211201\Models\SetApplicationSsoConfigRequest;
 
 use AlibabaCloud\SDK\Eiam\V20211201\Models\SetApplicationSsoConfigRequest\samlSsoConfig\attributeStatements;
+use AlibabaCloud\SDK\Eiam\V20211201\Models\SetApplicationSsoConfigRequest\samlSsoConfig\optionalRelayStates;
 use AlibabaCloud\Tea\Model;
 
 class samlSsoConfig extends Model
@@ -66,6 +67,11 @@ class samlSsoConfig extends Model
     public $nameIdValueExpression;
 
     /**
+     * @var optionalRelayStates[]
+     */
+    public $optionalRelayStates;
+
+    /**
      * @description Specifies whether to calculate the signature for the response. You cannot set ResponseSigned and AssertionSigned to false at the same time.
      *
      *   true
@@ -115,6 +121,7 @@ class samlSsoConfig extends Model
         'idPEntityId'           => 'IdPEntityId',
         'nameIdFormat'          => 'NameIdFormat',
         'nameIdValueExpression' => 'NameIdValueExpression',
+        'optionalRelayStates'   => 'OptionalRelayStates',
         'responseSigned'        => 'ResponseSigned',
         'signatureAlgorithm'    => 'SignatureAlgorithm',
         'spEntityId'            => 'SpEntityId',
@@ -151,6 +158,15 @@ class samlSsoConfig extends Model
         }
         if (null !== $this->nameIdValueExpression) {
             $res['NameIdValueExpression'] = $this->nameIdValueExpression;
+        }
+        if (null !== $this->optionalRelayStates) {
+            $res['OptionalRelayStates'] = [];
+            if (null !== $this->optionalRelayStates && \is_array($this->optionalRelayStates)) {
+                $n = 0;
+                foreach ($this->optionalRelayStates as $item) {
+                    $res['OptionalRelayStates'][$n++] = null !== $item ? $item->toMap() : $item;
+                }
+            }
         }
         if (null !== $this->responseSigned) {
             $res['ResponseSigned'] = $this->responseSigned;
@@ -199,6 +215,15 @@ class samlSsoConfig extends Model
         }
         if (isset($map['NameIdValueExpression'])) {
             $model->nameIdValueExpression = $map['NameIdValueExpression'];
+        }
+        if (isset($map['OptionalRelayStates'])) {
+            if (!empty($map['OptionalRelayStates'])) {
+                $model->optionalRelayStates = [];
+                $n                          = 0;
+                foreach ($map['OptionalRelayStates'] as $item) {
+                    $model->optionalRelayStates[$n++] = null !== $item ? optionalRelayStates::fromMap($item) : $item;
+                }
+            }
         }
         if (isset($map['ResponseSigned'])) {
             $model->responseSigned = $map['ResponseSigned'];